sales-agent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
openai==1.3.5
requests==2.31.0
python-multipart==0.0.6
alembic==1.12.1
httpx==0.25.1
aiosmtplib==3.0.1
email-validator==2.1.0
